# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BattleReadyShelf API
  version: 0.1.0
  description: BattleReadyShelf API
servers:
  - url: /
paths:
...
  /api/v1/users/me:
    get:
      tags: [ "Users" ]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/my/wargear-definitions:
    get:
      tags: [ "PersonalWargearDefinitions" ]
      summary: List the signed-in user's own wargear
      description: >-
        Both kinds of personal wargear: customisations of shared wargear (which carry a
        baseWargearDefinitionId) and wargear the user named themselves, including anything created
        implicitly by naming wargear on a personal model definition.
      operationId: getMyWargearDefinitions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: This user's wargear
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WargearDefinition'
    post:
      tags: [ "PersonalWargearDefinitions" ]
      summary: Create wargear of the user's own
      operationId: createMyWargearDefinition
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWargearDefinitionRequest'
      responses:
        '201':
          description: The created wargear
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WargearDefinition'
        '409':
          description: The user already has wargear with that name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/wargear-definitions/shared:
    get:
      tags: [ "PersonalWargearDefinitions" ]
      summary: List the shared wargear on its own
      description: >-
        GET /api/v1/wargear-definitions returns the shared catalogue and the user's own together.
        This endpoint returns the shared wargear alone: to offer it for customisation, and to diff a
        personal copy against the wargear it was forked from.
      operationId: getSharedWargearDefinitions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The shared wargear
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WargearDefinition'
  /api/v1/my/wargear-definitions/{wargearDefinitionId}:
    put:
      tags: [ "PersonalWargearDefinitions" ]
      summary: Rename one of the user's own wargear definitions
      operationId: updateMyWargearDefinition
      security:
        - bearerAuth: [ ]
      parameters:
        - name: wargearDefinitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWargearDefinitionRequest'
      responses:
        '200':
          description: The updated wargear
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WargearDefinition'
        '404':
          description: No such wargear belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The user already has wargear with that name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags: [ "PersonalWargearDefinitions" ]
      summary: Delete one of the user's own wargear definitions
      operationId: deleteMyWargearDefinition
      security:
        - bearerAuth: [ ]
      parameters:
        - name: wargearDefinitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The wargear was deleted
        '404':
          description: No such wargear belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Model definition options still reference this wargear
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/wargear-definitions/from/{wargearDefinitionId}:
    post:
      tags: [ "PersonalWargearDefinitions" ]
      summary: Start customising a shared wargear definition
      description: >-
        Copies the shared wargear into one owned by the caller, which they can then rename and
        attach to their own model definitions. Idempotent - returns the existing customisation if
        they already have one.
      operationId: customiseWargearDefinition
      security:
        - bearerAuth: [ ]
      parameters:
        - name: wargearDefinitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The personal copy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WargearDefinition'
        '400':
          description: The given wargear is not shared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Wargear definition not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/my/paints:
    get:
      tags: [ "PersonalPaints" ]
      summary: List the signed-in user's own paints
      description: >-
        Both kinds of personal paint: customisations of catalogue paints (which carry a basePaintId)
        and paints the user added themselves, such as a mix of their own or a brand the catalogue
        doesn't carry.
      operationId: getMyPaints
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: This user's paints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Paint'
    post:
      tags: [ "PersonalPaints" ]
      summary: Add a paint of the user's own
      operationId: createMyPaint
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePaintRequest'
      responses:
        '201':
          description: The created paint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paint'
        '409':
          description: The user already has a paint with that name and brand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/paints/shared:
    get:
      tags: [ "PersonalPaints" ]
      summary: List the shared paint catalogue on its own
      description: >-
        GET /api/v1/paints returns the shared catalogue and the user's own together. This endpoint
        returns the shared paints alone: to offer them for customisation, and to diff a personal
        copy against the paint it was forked from.
      operationId: getSharedPaints
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The shared paints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Paint'
  /api/v1/my/paints/{paintId}:
    put:
      tags: [ "PersonalPaints" ]
      summary: Update one of the user's own paints
      operationId: updateMyPaint
      security:
        - bearerAuth: [ ]
      parameters:
        - name: paintId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePaintRequest'
      responses:
        '200':
          description: The updated paint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paint'
        '404':
          description: No such paint belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The user already has a paint with that name and brand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags: [ "PersonalPaints" ]
      summary: Delete one of the user's own paints
      description: For a customisation this is the "revert to the catalogue version" action.
      operationId: deleteMyPaint
      security:
        - bearerAuth: [ ]
      parameters:
        - name: paintId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The paint was deleted
        '404':
          description: No such paint belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Paint recipes still reference this paint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/paints/from/{paintId}:
    post:
      tags: [ "PersonalPaints" ]
      summary: Start customising a shared paint
      description: >-
        Copies the catalogue paint into one owned by the caller, which they can then rename or
        recolour - useful when a paint has been reformulated, or the user thins or mixes it and
        wants their own swatch. Idempotent - returns the existing customisation if they already
        have one.
      operationId: customisePaint
      security:
        - bearerAuth: [ ]
      parameters:
        - name: paintId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The personal copy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paint'
        '400':
          description: The given paint is not shared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Paint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
      required: [ message ]
...
    WargearDefinition:
      type: object
      description: >-
        A canonical piece of wargear shared by every model definition that can be equipped with it,
        so it is named once rather than copied per model. Ids form a single global namespace
        because the reference dataset already reuses them across factions.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalId:
          type: string
          readOnly: true
          description: Stable identifier from the reference dataset; null for wargear created by hand.
        ownerUserId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            Null for the shared catalogue. Set when a user named wargear the catalogue doesn't
            have while building a personal model definition, in which case only they can see it.
        name:
          type: string
        baseWargearDefinitionId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            The shared wargear this personal one was forked from, so the app can show how it
            differs and offer to revert. Null both for shared wargear and for wargear a user
            named themselves.
        usageCount:
          type: integer
          format: int32
          readOnly: true
          description: How many published model definitions currently reference this wargear.
        hidden:
          type: boolean
          readOnly: true
          description: >-
            Whether the caller has taken this out of their own pickers. Per-user, so hiding shared
            wargear affects nobody else. Only ever true on the /api/v1/my lists, which return
            hidden definitions so they can be shown again; the lists behind pickers omit them.
      required: [ name ]
    UpdateWargearDefinitionRequest:
      type: object
      properties:
        name:
          type: string
      required: [ name ]
...
    PaintType:
      type: string
      description: >-
        What the paint is for, which is also roughly the order it goes on the model. Recipes are
        sorted by the step the user gives them rather than by this, but it makes a long catalogue
        searchable and tells you why two similar colours are not interchangeable.
      enum: [ BASE, LAYER, SHADE, CONTRAST, DRY, TECHNICAL, PRIMER, VARNISH, OTHER ]
    Paint:
      type: object
      description: >-
        A paint that can be recorded in a recipe, e.g. Citadel's "Leadbelcher". Brand is part of the
        identity because names are only unique within one: several manufacturers sell a "Bone
        White".
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalId:
          type: string
          readOnly: true
          nullable: true
          description: Stable identifier from the reference dataset; null for paints added by hand.
        ownerUserId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            Null for the shared catalogue. Set when a user added a paint of their own - a mix, or a
            brand the catalogue doesn't carry - in which case only they can see it.
        basePaintId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            The catalogue paint this personal one was forked from, so the app can show how it
            differs and offer to revert. Null both for catalogue paints and for paints a user added
            themselves.
        name:
          type: string
          description: The paint's name, e.g. "Leadbelcher".
        brand:
          type: string
          nullable: true
          description: Who makes it, e.g. "Citadel". Null for a user's own mix.
        paintType:
          $ref: '#/components/schemas/PaintType'
        hexColour:
          type: string
          nullable: true
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: >-
            A swatch colour as #RRGGBB. An indicator rather than a match - metallics, shades and
            contrast paints cannot be represented honestly by a flat colour.
        usageCount:
          type: integer
          format: int32
          readOnly: true
          description: How many paint recipes currently reference this paint.
        hidden:
          type: boolean
          readOnly: true
          description: >-
            Whether the caller has taken this out of their own pickers. Per-user, so hiding a
            catalogue paint affects nobody else. Only ever true on the /api/v1/my lists, which
            return hidden definitions so they can be shown again; the lists behind pickers omit them.
      required: [ name ]
...
    UpdatePaintRequest:
      type: object
      description: >-
        The complete proposed state of a paint. An omitted or null brand, paintType or hexColour
        proposes clearing that field rather than leaving it alone.
      properties:
        name:
          type: string
        brand:
          type: string
          nullable: true
        paintType:
          $ref: '#/components/schemas/PaintType'
        hexColour:
          type: string
          nullable: true
          pattern: '^#[0-9A-Fa-f]{6}$'
      required: [ name ]
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/UserRole'
        themePreference:
          allOf:
            - $ref: '#/components/schemas/ThemePreference'
      required: [ id, email ]
    UserRole:
      type: string
      enum: [ GUEST, USER, ADMIN, SUPERADMIN ]
    ThemePreference:
      type: string
      description: >-
        A user's preferred UI colour scheme. AUTO follows the device's system preference.
      enum: [ LIGHT, DARK, AUTO ]
...